+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [1.0.0-rc.6] - 2019-02-25
+### Fixed
+- Closing `CreateSaleForm` after submitting
+- Using asset records instead of asset codes in `CreateSaleForm`
+- Populating `CreateSaleForm`
+- Using `saleEnumType` for sale type
+
 ## [1.0.0-rc.5] - 2019-02-25
 ### Fixed
 - Add loadKvEntriesAccountRoleIds after sign in
@@ -136,7 +143,8 @@ for further information about branching and tagging conventions.
 
 ## [1.0.0-rc.0] - 2019-02-18
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.0.0-rc.5...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.0.0-rc.6...HEAD
+[1.0.0-rc.6]: https://github.com/tokend/web-client/compare/1.0.0-rc.5...1.0.0-rc.6
 [1.0.0-rc.5]: https://github.com/tokend/web-client/compare/1.0.0-rc.4...1.0.0-rc.5
 [1.0.0-rc.4]: https://github.com/tokend/web-client/compare/1.0.0-rc.3...1.0.0-rc.4
 [1.0.0-rc.3]: https://github.com/tokend/web-client/compare/1.0.0-rc.2...1.0.0-rc.3